refactor: Remove unused imports and commented-out code for improved clarity and maintainability; add new API route for score evaluation
