+++ app/analytics/components/recent-logs.tsx
@@ -1,4 +1,3 @@
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Skeleton } from "@/components/ui/skeleton";
 
 export function RecentLogs() {
@@ -6,12 +5,7 @@ export function RecentLogs() {
     <div className="space-y-8">
       {Array.from({ length: 5 }).map((_, i) => (
         <div key={i} className="flex items-center space-x-4">
-          <Skeleton className="flex-shrink-0 h-9 w-9 rounded-full">
-            {/* <Avatar className="h-9 w-9">
-              <AvatarImage src="/avatars/01.png" alt="Avatar" />
-              <AvatarFallback>OM</AvatarFallback>
-            </Avatar> */}
-          </Skeleton>
+          <Skeleton className="flex-shrink-0 h-9 w-9 rounded-full"></Skeleton>
           <div className="ml-4 space-y-1 w-full">
             <div className="text-sm font-medium leading-none">
               <Skeleton className="h-4 w-full"></Skeleton>
