+++ app/playground/page.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { cubicBezier, motion } from "framer-motion";
 import Link from "next/link";
 import { Card1, Card2, Card3 } from "@/components/sections/tools";
