+++ app/playground/tweets/page.tsx
@@ -32,11 +32,6 @@ import { useMemo } from "react";
 
 import { useEvaluationDialog } from "@/hooks/use-evaluation-dialog";
 
-// export const metadata: Metadata = {
-//   title: "Playground",
-//   description: "The OpenAI Playground built using the components.",
-// };
-
 export default function PlaygroundPage() {
   const methods = useForm();
   const { isSidebarOpen, openSidebar, closeSidebar, toggleSidebar } =
